First commit
This is the initial form of the website (just one page). I need to optimize a few things and this is where the website is functional completely so I just want to upload it here for safekeeping if I end up screwing stuff up and have the need to revert to a backup. I will try to accomplish this specific look, but just in a more "responsive design" way. After that I will commit changes again and show what I changed.