Initial Commit for test website (navbar dropdown)
This commit shows what I am doing to the navbar. I want a dropdown for the contact button and it is difficult to do that, so I am just trying to find out the inner workings of it so I can understand how to replicate it over on the actual website.