Commit for Functionality
Going to commit this for reference later. Everything works its just it spams the event to console. I need to fix this with some logic so it doesn't do anything without checking what is going on with the other hover methods and boxes.

Meaning, it will not change what is going on until the situation is true (cursor is not hovering over boxes or it is hovering over specified boxes)