Update for 2/28/22
After meeting with Diego we discussed some changes, this commit has a few of the changes that I published live in front of him. This commit is not signifying the end of my coding session, it is just to upload a recent copy of the website for Diego to view.