Going to make this an issue
Bug need to add. The back button in the coming soon document doesn't work, it always returns to home page. I don't know why, the javascript seems to be running correctly. However, this isn't an important issue so I will fix this later. For now I will just log it in bugs and I will, fix this in version 2.