Fixed everything with the selection tool for pictures
Going to update this so Diego can see the changes. Going to figure out what to do with the box that shows the picture (might do some media query flex box stuff). I feel like that would be easy and it would solve my problem. I also think I need to implement a zoom in feature for the pictures. I might do that in version 2 tho.